+++ lib/src/constants/boolean.dart
@@ -1,0 +1,6 @@
+// ignore_for_file: constant_identifier_names
+abstract class BooleanConst {
+  const BooleanConst._();
+  static const bool TRUE = true;
+  static const bool FALSE = false;
+}