+++ lib/src/extensions/context_extension.dart
@@ -9,12 +9,9 @@ extension BuildContextX on BuildContext {
   ColorScheme get colorScheme => Theme.of(this).colorScheme;
 
   /// check isDarkTheme enabled
-  bool get isDarkTheme =>
-      theme.brightness == Brightness.dark ||
-      colorScheme.brightness == Brightness.dark;
+  bool get isDarkTheme => theme.brightness == Brightness.dark || colorScheme.brightness == Brightness.dark;
 
-  Color themedColor(Color light, [Color? dark]) =>
-      isDarkTheme ? (dark ?? light) : light;
+  Color themedColor(Color light, [Color? dark]) => isDarkTheme ? (dark ?? light) : light;
 
   /// get textTheme
   TextTheme get textTheme => Theme.of(this).textTheme;
@@ -116,12 +113,10 @@ extension BuildContextX on BuildContext {
   MediaQueryData get mq => MediaQuery.of(this);
 
   /// screen width
-  double width([bool responsive = false]) =>
-      responsive ? SizeUtils.instance.width : size.width;
+  double get width => size.width;
 
   /// screen height
-  double height([bool responsive = false]) =>
-      responsive ? SizeUtils.instance.height : size.height;
+  double get height => size.height;
 
   /// window padding
   /// The parts of the display that are partially obscured by system UI, typically by the hardware display "notches" or the system status bar.
@@ -141,12 +136,12 @@ extension BuildContextX on BuildContext {
 
   ///percent with
   double pw(double percent, [bool responsive = false]) {
-    return width(responsive) * percent;
+    return width * percent;
   }
 
   ///percent height
   double ph(double percent, [bool responsive = false]) {
-    return height(responsive) * percent;
+    return height * percent;
   }
 
   /// remove keyboard focus
@@ -155,8 +150,7 @@ extension BuildContextX on BuildContext {
   }
 
   /// check whether keyboard has focus or not
-  bool get hasFocus =>
-      FocusScope.of(this).hasFocus || FocusScope.of(this).hasPrimaryFocus;
+  bool get hasFocus => FocusScope.of(this).hasFocus || FocusScope.of(this).hasPrimaryFocus;
 
   /// extension to get responsive value according to device type [ mobile,tablet,desktop ]
   T responsiveValue<T>(