+++ lib/src/extensions/num_extension.dart
@@ -1,14 +1,14 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:fp_util/src/responsive/size_utils.dart';
 
 extension NumX on num {
   /// horizontal spacing widget
-  SizedBox horizontalSpace([bool responsive = false]) =>
-      SizedBox(width: responsive ? toDouble().w : toDouble());
+  SizedBox get horizontalSpace => SizedBox(width: toDouble());
 
   /// vertical spacing widget
-  SizedBox verticalSpace([bool responsive = false]) =>
-      SizedBox(height: responsive ? toDouble().h : toDouble());
+  SizedBox get verticalSpace => SizedBox(height: toDouble());
 
   /// duration in microseconds
   Duration get microseconds => Duration(microseconds: toInt());
@@ -29,67 +29,90 @@ extension NumX on num {
   Duration get days => Duration(days: toInt());
 
   /// get delayed future in microsecond
-  Future get delayedMicroSeconds async => Future.delayed(toInt().microseconds);
+  Future delayedMicroSeconds([
+    FutureOr<dynamic> Function()? callback,
+  ]) {
+    return Future.delayed(
+      toInt().microseconds,
+      callback,
+    );
+  }
 
   /// get delayed future in millisecond
-  Future get delayedMilliSeconds async => Future.delayed(toInt().milliseconds);
+  Future delayedMilliSeconds([
+    FutureOr<dynamic> Function()? callback,
+  ]) {
+    return Future.delayed(
+      toInt().milliseconds,
+      callback,
+    );
+  }
 
   /// get delayed future in seconds
-  Future get delayedSeconds async => Future.delayed(toInt().seconds);
+  Future delayedSeconds([
+    FutureOr<dynamic> Function()? callback,
+  ]) {
+    return Future.delayed(
+      toInt().seconds,
+      callback,
+    );
+  }
 
   /// get delayed future in minutes
-  Future get delayedMinutes async => Future.delayed(toInt().minutes);
+  Future delayedMinutes([
+    FutureOr<dynamic> Function()? callback,
+  ]) {
+    return Future.delayed(
+      toInt().minutes,
+      callback,
+    );
+  }
 
   /// get delayed future in hours
-  Future get delayedHours async => Future.delayed(toInt().hours);
+  Future delayedHours(
+    FutureOr<dynamic> Function()? callback,
+  ) {
+    return Future.delayed(
+      toInt().hours,
+      callback,
+    );
+  }
 
   /// get circular radius
-  Radius circularRadius([bool responsive = false]) =>
-      Radius.circular(responsive ? toDouble().px : toDouble());
+  Radius get circularRadius => Radius.circular(toDouble());
 
   /// get circular border radius geometry
-  BorderRadiusGeometry circular([bool responsive = false]) =>
-      BorderRadius.circular(responsive ? toDouble().px : toDouble());
+  BorderRadiusGeometry get circular => BorderRadius.circular(toDouble());
 
   /// circular border radius
-  BorderRadius circularBorderRadius([bool responsive = false]) =>
-      BorderRadius.circular(responsive ? toDouble().px : toDouble());
+  BorderRadius get circularBorderRadius => BorderRadius.circular(toDouble());
 
   /// get all EdgeInsets
-  EdgeInsets all([bool responsive = false]) =>
-      EdgeInsets.all(responsive ? toDouble().px : toDouble());
+  EdgeInsets get all => EdgeInsets.all(toDouble());
 
   /// get left EdgeInsets
-  EdgeInsets padLeft([bool responsive = false]) =>
-      EdgeInsets.only(left: responsive ? toDouble().w : toDouble());
+  EdgeInsets get padLeft => EdgeInsets.only(left: toDouble());
 
   /// get top EdgeInsets
-  EdgeInsets padTop([bool responsive = false]) =>
-      EdgeInsets.only(top: responsive ? toDouble().h : toDouble());
+  EdgeInsets get padTop => EdgeInsets.only(top: toDouble());
 
   /// get right EdgeInsets
-  EdgeInsets padRight([bool responsive = false]) =>
-      EdgeInsets.only(right: responsive ? toDouble().w : toDouble());
+  EdgeInsets get padRight => EdgeInsets.only(right: toDouble());
 
   /// get bottom EdgeInsets
-  EdgeInsets padBottom([bool responsive = false]) =>
-      EdgeInsets.only(bottom: responsive ? toDouble().h : toDouble());
+  EdgeInsets get padBottom => EdgeInsets.only(bottom: toDouble());
 
   /// horizontal spacing for edge-inset
-  EdgeInsets padX([bool responsive = false]) =>
-      EdgeInsets.symmetric(horizontal: responsive ? toDouble().w : toDouble());
+  EdgeInsets get padX => EdgeInsets.symmetric(horizontal: toDouble());
 
   /// vertical spacing for edge-inset
-  EdgeInsets padY([bool responsive = false]) =>
-      EdgeInsets.symmetric(vertical: responsive ? toDouble().h : toDouble());
+  EdgeInsets get padY => EdgeInsets.symmetric(vertical: toDouble());
 
   /// convert spacing widget into sliver widget
-  SliverToBoxAdapter sliverHorizontalSpace([bool responsive = false]) =>
-      SliverToBoxAdapter(child: horizontalSpace(responsive));
+  SliverToBoxAdapter get sliverHorizontalSpace => SliverToBoxAdapter(child: horizontalSpace);
 
   /// convert horizontal spacing widget into sliver widget
-  SliverToBoxAdapter sliverVerticalSpace([bool responsive = false]) =>
-      SliverToBoxAdapter(child: verticalSpace(responsive));
+  SliverToBoxAdapter get sliverVerticalSpace => SliverToBoxAdapter(child: verticalSpace);
 
   /// responsive px
   double get px => SizeUtils.instance.getSize(toDouble());